Day 35: refactor to half precision + profling row-wise quantization vs bnb in python

+++ quantization/src/LLM_int8.cuh
@@ -1,6 +1,7 @@
 #pragma once
+#include <cuda_fp16.h>
 
-void rowWiseQuant8bits_gpu(float *X, size_t nRows, size_t nCols, int8_t *q_X);
+void rowWiseQuant8bits_gpu(__half *X, size_t nRows, size_t nCols, int8_t *q_X);
 void columnWiseQuant8bits_gpu(float *X, size_t nRows, size_t nCols, int8_t *q_X);
 
 template <typename IN_TYPE, typename OUT_TYPE>
